requirements.txt
discord.py==2.3.2
python-binance==1.0.19
python-dotenv==1.0.0
ccxt==4.1.92

